feat(ArchivePage): integrate page with external API
